requirements.txt
EbookLib==0.17
beautifulsoup4==4.12.2
nltk==3.8.1
tqdm==4.66.1
openai
python-dotenv==1.0.0
jsonlines==3.1.0 